优化上传下载

+++ family-disk-service/family-disk-service-file/src/main/java/com/jflove/file/impl/FileServiceImpl.java
@@ -131,8 +131,6 @@ public class FileServiceImpl implements IFileService {
         };
     }
 
-
-
     @Override
     public StreamObserver<FileTransmissionDTO> writeByte(StreamObserver<FileTransmissionRepDTO> response) {
         return new StreamObserver<FileTransmissionDTO>() {
@@ -141,17 +139,30 @@ public class FileServiceImpl implements IFileService {
             public void onNext(FileTransmissionDTO data) {
                 log.info("dubbo文件传输流,接收到传输,文件名称:{},文件类型:{},文件总大小:{},当前分片起:{},当前分片止:{},分片数量:{},第{}分片",
                         data.getName(),data.getType(),data.getTotalSize(),data.getRangeStart(),data.getRangeEnd(),data.getShardingNum(),data.getShardingSort());
-                //判断这是不是最后一片
+                //这个分片写入前,已写入的分片
                 File[] fs = new File(tempPath).listFiles(e->e.getName().startsWith(data.getFileMd5()));
-                boolean ok = fs.length == data.getShardingNum() - 1;
+                if(fs.length == data.getShardingNum()){//磁盘中的分片数量已经是最大的分片,则不执行下面的逻辑,是某些极少数以外导致的,等第二天系统会自动清理昨天的临时文件,到时候就可以重新上传了
+                    response.onNext(new FileTransmissionRepDTO(data.getName(),data.getFileMd5(),true,"磁盘分片已全部写入,跳过重复写入"));
+                    return;
+                }
                 try {
                     Path p = Path.of(String.format("%s/%s-%s%s", tempPath,data.getFileMd5(), String.valueOf(data.getShardingSort()), tempFileSuffix));
-                    Files.deleteIfExists(p);//如果分片存在,先删除这个分片,重新写入
-                    Files.write(p,data.getShardingStream());
+                    if(Files.notExists(p)) {
+                        Files.write(p, data.getShardingStream());
+                    }
                 }catch (IOException e){
                     log.error("分片存储异常",e);
                     response.onNext(new FileTransmissionRepDTO(data.getName(),data.getFileMd5(),false,"分片文件存储异常"));
                     return;
+                }
+                //判断所有的分片是否都存在如果存在,则开始合并
+                boolean ok = true;
+                for (int i = 0; i < data.getShardingNum(); i++) {
+                    Path p = Path.of(String.format("%s/%s-%s%s", tempPath,data.getFileMd5(), String.valueOf(i), tempFileSuffix));
+                    if(Files.notExists(p)){
+                        ok = false;
+                        break;
+                    }
                 }
                 if(ok){//这已经是最后一片了,开始合并
                     try {
@@ -235,7 +246,10 @@ public class FileServiceImpl implements IFileService {
                 log.info("dubbo文件传输流,接收到传输,文件名称:{},文件类型:{},文件总大小:{},分片个数:{},当前分片序号:{},流大小:{},第{}分片",
                         data.getName(),data.getType(),data.getTotalSize(),data.getShardingNum(),data.getShardingSort(),data.getShardingStream().length,data.getShardingSort());
                 try {
-                    Files.write(Path.of(String.format("%s/%s-%s%s", tempPath, data.getFileMd5(), String.valueOf(data.getShardingSort()), tempFileSuffix)), data.getShardingStream());
+                    Path s = Path.of(String.format("%s/%s-%s%s", tempPath, data.getFileMd5(), String.valueOf(data.getShardingSort()), tempFileSuffix));
+                    if(Files.notExists(s)) {
+                        Files.write(s, data.getShardingStream());
+                    }
                 }catch (IOException e){
                     log.error("分片存储异常",e);
                     response.onNext(new FileTransmissionRepDTO(data.getName(),data.getFileMd5(),false,"分片文件存储异常"));
